tox-requirements.txt
pytest
pytest-django
pytest-pythonpath
tox
factory_boy